Quote root password env variable
